+++ modules/apps/27-interchain-accounts/host/types/params.go
@@ -10,6 +10,8 @@ import (
 const (
 	// DefaultHostEnabled is the default value for the host param (set to true)
 	DefaultHostEnabled = true
+	// Maximum length of the allowlist
+	MaxAllowListLength = 500
 )
 
 var (
@@ -50,6 +52,7 @@ func (p Params) Validate() error {
 	return nil
 }
 
+<<<<<<< HEAD
 // ParamSetPairs implements params.ParamSet
 func (p *Params) ParamSetPairs() paramtypes.ParamSetPairs {
 	return paramtypes.ParamSetPairs{
@@ -71,6 +74,15 @@ func validateAllowlist(i interface{}) error {
 	allowMsgs, ok := i.([]string)
 	if !ok {
 		return fmt.Errorf("invalid parameter type: %T", i)
+=======
+func validateAllowlist(allowMsgs []string) error {
+	if len(allowMsgs) > MaxAllowListLength {
+		return fmt.Errorf("allow list length must not exceed %d items", MaxAllowListLength)
+	}
+
+	if slices.Contains(allowMsgs, AllowAllHostMsgs) && len(allowMsgs) > 1 {
+		return fmt.Errorf("allow list must have only one element because the allow all host messages wildcard (%s) is present", AllowAllHostMsgs)
+>>>>>>> 478f4c60 (imp: check length of slices of messages (#6256))
 	}
 
 	for _, typeURL := range allowMsgs {