+++ modules/apps/27-interchain-accounts/host/types/params_test.go
@@ -11,4 +11,11 @@ import (
 func TestValidateParams(t *testing.T) {
 	require.NoError(t, types.DefaultParams().Validate())
 	require.NoError(t, types.NewParams(false, []string{}).Validate())
+<<<<<<< HEAD
+=======
+	require.Error(t, types.NewParams(true, []string{""}).Validate())
+	require.Error(t, types.NewParams(true, []string{" "}).Validate())
+	require.Error(t, types.NewParams(true, []string{"*", "/cosmos.bank.v1beta1.MsgSend"}).Validate())
+	require.Error(t, types.NewParams(true, make([]string, types.MaxAllowListLength+1)).Validate())
+>>>>>>> 478f4c60 (imp: check length of slices of messages (#6256))
 }